refactor(auth): extract password validation error handling to shared utility
Move password validation error extraction logic to a reusable utility function to avoid code duplication between registration and password change flows
