+++ apps/web/app/api/mutations/useChangePassword.ts
@@ -3,26 +3,12 @@ import { AxiosError } from "axios";
 import { useTranslation } from "react-i18next";
 
 import { useToast } from "~/components/ui/use-toast";
+import { extractPasswordValidationErrors } from "~/utils/PasswordValidationErrorHandler";
 
 import { ApiClient } from "../api-client";
 import { useCurrentUserSuspense } from "../queries/useCurrentUser";
 
 import type { ChangePasswordBody } from "../generated-api";
-
-interface PasswordValidationError {
-  type: number;
-  schema?: {
-    minLength?: number;
-    maxLength?: number;
-    pattern?: string;
-    errorMessage?: string;
-    type?: string;
-    allOf?: unknown[];
-  };
-  path: string;
-  value: unknown;
-  message: string;
-}
 
 type ChangePasswordOptions = {
   data: ChangePasswordBody;
@@ -49,24 +35,12 @@ export function useChangePassword() {
     },
     onError: (error) => {
       if (error instanceof AxiosError) {
-        const errors = error.response?.data.errors || [];
-        const errorMessages: string[] = [];
-
-        errors.forEach((err: PasswordValidationError) => {
-          if (err.schema?.errorMessage) {
-            errorMessages.push(err.schema.errorMessage);
-          }
-        });
-
-        const uniqueMessages = [...new Set(errorMessages)];
-        const errorString = uniqueMessages.join("\n");
-
+        const errorMessage = extractPasswordValidationErrors(error);
         return toast({
           variant: "destructive",
-          description: errorString || error.response?.data.message,
+          description: errorMessage,
         });
       }
-
       toast({
         variant: "destructive",
         description: error.message,
