+++ apps/web/app/api/mutations/useRegisterUser.ts
@@ -3,6 +3,7 @@ import { useMutation } from "@tanstack/react-query";
 import { AxiosError } from "axios";
 
 import { useToast } from "~/components/ui/use-toast";
+import { extractPasswordValidationErrors } from "~/utils/PasswordValidationErrorHandler";
 
 import { ApiClient } from "../api-client";
 
@@ -26,9 +27,10 @@ export function useRegisterUser() {
     },
     onError: (error) => {
       if (error instanceof AxiosError) {
+        const errorMessage = extractPasswordValidationErrors(error);
         return toast({
           variant: "destructive",
-          description: error.response?.data.message,
+          description: errorMessage,
         });
       }
       toast({
